% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omim.R
\name{get_omim}
\alias{get_omim}
\title{Query OMIM server}
\usage{
get_omim(omim_id = 100100, text = FALSE, existflags = FALSE,
  allelicVariantList = FALSE, clinicalSynopsis = FALSE, seeAlso = FALSE,
  referenceList = FALSE, geneMap = FALSE, externalLinks = FALSE,
  contributors = FALSE, creationDate = FALSE, editHistory = FALSE,
  dates = FALSE, all = FALSE)
}
\arguments{
\item{omim_id}{OMIM ID. Default is 100100.}

\item{text}{Includes the text field sections with the entry. Default is FALSE.}

\item{existflags}{Include the 'exists' flags with the entry (clinical synopsis, allelic variant, gene map & phenotype map). Default is FALSE.}

\item{allelicVariantList}{Includes the allelic variant list with the entry. Default is FALSE.}

\item{clinicalSynopsis}{Include the clinical synopsis with the entry. Default is FALSE.}

\item{seeAlso}{Includes the 'see also' field with the entry. Default is FALSE.}

\item{referenceList}{Include the reference list with the entry. Default is FALSE.}

\item{geneMap}{Include the gene map/phenotype map data with the entry. Default is FALSE.}

\item{externalLinks}{Include the external links with the entry. Default is FALSE.}

\item{contributors}{Includes the 'contributors' field with the entry. Default is FALSE.}

\item{creationDate}{Includes the 'creation date' field with the entry. Default is FALSE.}

\item{editHistory}{Includes the 'edit history' field with the entry. Default is FALSE.}

\item{dates}{Include the dates data with the entry. Default is FALSE.}

\item{all}{the above data with the entry. Default is FALSE.}
}
\description{
This function builds the URL and queries the OMIM server
}
\examples{
get_omim(303600, geneMap = TRUE)
}
\keyword{ID}
\keyword{OMIM}
