% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pheno_key.R
\name{get_pheno_key}
\alias{get_pheno_key}
\title{Get phenotype mapping key for gene/s associated with an OMIM entry}
\usage{
get_pheno_key(xml)
}
\arguments{
\item{xml}{XML result from get_omim()}
}
\description{
Returns the phenotype mapping key of gene/s associated with an OMIM entry.
Assumes that the phenotype mapping key is stored in the XML node /omim/entryList/entry/phenotypeMapList/phenotypeMap/phenotypeMappingKey.
}
\examples{
get_pheno_key(get_omim(303600, geneMap=TRUE))
}
\keyword{key}
\keyword{mapping}
\keyword{phenotype}
