% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_title.R
\name{get_title}
\alias{get_title}
\title{Get title from an OMIM ID}
\usage{
get_title(xml)
}
\arguments{
\item{xml}{XML result from get_omim()}
}
\description{
This function parses the XML result and extracts the title of an OMIM ID
}
\examples{
get_title(get_omim(100200))
}
\keyword{title}
