% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypic_series.R
\name{phenotypic_series}
\alias{phenotypic_series}
\title{OMIM Phenotypic Series}
\usage{
phenotypic_series(series = "PS163950")
}
\arguments{
\item{series}{Phenotypic Series Number. Default is PS163950.}
}
\description{
This function returns OMIM IDs given a OMIM Phenotypic Series ID.
}
\examples{
phenotypic_series(series = 'PS308350')
}
\keyword{Number}
\keyword{Phenotypic}
\keyword{Series}
