% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_phenotype.R
\name{search_phenotype}
\alias{search_phenotype}
\title{Search OMIM}
\usage{
search_phenotype(search_term = "noonan")
}
\arguments{
\item{series}{Search term. Default is "noonan".}
}
\description{
This function returns OMIM phenotypes given a match of a search term in the text meta field of OMIM phenotype entries.
}
\examples{
search_phenotype(search_term = 'noonan')
}
\keyword{OMIM}
\keyword{Search}
