% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_key.R
\name{set_key}
\alias{set_key}
\title{Set API key}
\usage{
set_key(key)
}
\arguments{
\item{key}{Your API key}
}
\description{
This function sets the API key
}
\examples{
set_key('aAS5ASdf35jasdnadsfj')
}
\keyword{key}

