% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_to_omim.R
\name{gene_to_omim}
\alias{gene_to_omim}
\title{Gene search}
\usage{
gene_to_omim(gene_symbol = "FGFR2")
}
\arguments{
\item{gene_symbol}{Gene Symbol. Default is FGFR2.}
}
\description{
This function builds the URL to search the OMIM server based on an input gene
}
\examples{
gene_to_omim(TTN)
}
\keyword{Gene}
\keyword{Symbol}
