% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene.R
\name{get_gene}
\alias{get_gene}
\title{Get gene/s from an OMIM ID}
\usage{
get_gene(xml)
}
\arguments{
\item{xml}{XML result from get_omim()}
}
\description{
This function parses the XML result and extracts the genes associated with an OMIM ID
}
\examples{
get_gene(get_omim(303600, geneMap=TRUE))
}
\keyword{gene}

