% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inheritance.R
\name{get_inheritance}
\alias{get_inheritance}
\title{Get inheritance model from an OMIM ID}
\usage{
get_inheritance(xml)
}
\arguments{
\item{xml}{XML result from get_omim()}
}
\description{
This function parses the XML result and extracts the inheritance model associated with an OMIM ID
}
\examples{
get_inheritance(get_omim(303600, geneMap=TRUE, clinicalSynopsis=TRUE))
}
\keyword{inheritance}

