% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pheno_key.R
\name{get_pheno_key}
\alias{get_pheno_key}
\title{Get phenotype mapping key for an OMIM gene}
\usage{
get_pheno_key(xml)
}
\arguments{
\item{xml}{XML result from get_omim()}
}
\description{
This function parses the XML result and extracts the phenotype mapping key for the gene associated with an OMIM ID
}
\examples{
get_pheno_key(get_omim(303600, geneMap=TRUE))
}
\keyword{key}
\keyword{mapping}
\keyword{phenotype}
