% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypic_series.R
\name{phenotypic_series}
\alias{phenotypic_series}
\title{OMIM Phenotypic Series}
\usage{
phenotypic_series(series = "PS163950", show_query = FALSE)
}
\arguments{
\item{series}{Phenotypic Series Number. Default is PS163950.}

\item{show_query}{Show API query. Default is FALSE.}
}
\description{
This function returns OMIM IDs given a OMIM Phenotypic Series ID.
Recursive apply (rapply) and grep is used because phenotypeMimNumber is not always present in the same XML structure.
}
\examples{
phenotypic_series(series = 'PS308350')
}
\keyword{Number}
\keyword{Phenotypic}
\keyword{Series}
