% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_phenotype.R
\name{search_phenotype}
\alias{search_phenotype}
\title{Search OMIM}
\usage{
search_phenotype(search_term = "noonan syndrome", show_query = FALSE)
}
\arguments{
\item{show_query}{Show API query. Default is FALSE.}

\item{series}{Search term. Default is "noonan syndrome".}
}
\description{
This function returns OMIM phenotypes given a match of a search term in the text meta field of OMIM phenotype entries.
Recursive apply (rapply) and grep is used because phenotypeMimNumber is not always present in the same XML structure.
}
\examples{
search_phenotype(search_term = 'noonan syndrome')
}
\keyword{OMIM}
\keyword{Search}
